backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
redis==5.0.1
websockets==12.0
httpx==0.26.0
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
python-dotenv==1.0.0
